+++ core/domain/interface/order/complex_order.go
@@ -88,17 +88,17 @@ type (
 		// 商品图片
 		MainImage string `db:"image"`
 		// 商品单价
-		Price float32 `db:"-"`
+		Price int64 `db:"-"`
 		// 商品实际单价
-		FinalPrice float32 `db:"-"`
+		FinalPrice int64 `db:"-"`
 		// 数量
 		Quantity int32 `db:"quantity"`
 		// 退回数量(退货)
 		ReturnQuantity int32 `db:"return_quantity"`
 		// 金额
-		Amount float64 `db:"amount"`
+		Amount int64 `db:"amount"`
 		// 最终金额, 可能会有优惠均摊抵扣的金额
-		FinalAmount float64 `db:"final_amount"`
+		FinalAmount int64 `db:"final_amount"`
 		// 是否发货
 		IsShipped int32 `db:"is_shipped"`
 		// 其他数据
